requirements.txt
## Python 3.9 기준

Django==4.1.2
django-ninja==0.20.0
django-storages==1.13.2
django-cors-headers==4.2.0
psycopg2-binary==2.9.9
boto3==1.28.35
requests==2.31.0
uvicorn==0.23.2
beautifulsoup4==4.12.2
openpyxl==3.1.2
celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1
redis==5.0.0
